Fix the versioning & publishing scripts
These were the steps that were previously being taken during the version process:

1. Run `npm test && npm build`
2. Bump version in `package.json` (automatically done by npm)
3. Create the “Update dist for release” commit
4. Create a `release` branch and force-add the `dist` folder
5. Commit and tag the version (automatically done by npm)
6. Push the tags, switch to `master`, delete the `release` branch, and push the commits
7. Publish to npm

This resulted in a “Update dist for release” commit that included just the changed version number, while the commit created by npm would have the `dist` folder. The tagged release would have the `dist` folder but the version on npm didn’t.

Here are the new steps introduced by this commit:

1. Run `npm test && npm build`
2. Bump version in `package.json` (automatically done by npm)
3. Create a “Update version number” commit
4. Create a `release` branch and force-add the `dist` folder
5. Commit and tag the version (automatically done by npm)
6. Publish to npm
7. Push the tags, switch to `master`, delete the `release` branch, and push the commits

This results in both the tagged release and the npm package having the `dist` folder.